pyproject.toml
[tool.poetry]
name = "customer-support-assistant"
version = "0.1.0"
description = "A customer support assistant backend with AI integration"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
sqlalchemy = "^2.0.25"
pydantic = "^2.6.0"
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-multipart = "^0.0.6"
aiohttp = "^3.9.3"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.1"
python-dotenv = "^1.0.0"
groq = "^0.4.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.5"
black = "^24.1.1"
isort = "^5.13.2"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
multi_line_output = 3 